refactor(set_guest_user_password): move set_guest_user_password from core to auth
simplify cryptography import statement
update scheduled_tasks.py accordingly
